+++ Client/src/components/layout/ProfilePage.tsx/ProfilePageLayout.tsx
@@ -1,15 +1,19 @@
 import { ReactNode } from "react";
 import ProfilePageHeader from "./ProfilePageHeader";
 
+import DragDropContextWrapper from "../DragDropContxtWrapper";
+
 type ProfilePageLayoutProps = {
   children: ReactNode;
 };
 const ProfilePageLayout = ({ children }: ProfilePageLayoutProps) => {
   return (
-    <div className="dark:bg-slate-800 min-h-screen">
-      <ProfilePageHeader />
-      {children}
-    </div>
+    <DragDropContextWrapper>
+      <div className="dark:bg-slate-800 min-h-screen">
+        <ProfilePageHeader />
+        {children}
+      </div>
+    </DragDropContextWrapper>
   );
 };
